Поправлена верстка страниц, перенесено в один файл
