+++ packages/client/src/app.tsx
@@ -10,7 +10,7 @@ import { TestPage } from 'pages/test';
 import { ProtectedRoutes } from 'utils/protected-routes';
 import { Profile } from 'pages/profile';
 import { MainPage } from 'pages/main';
-import { LoginPage, RegisterPage } from 'pages/register-login';
+import { LoginPage, RegisterPage } from 'pages/auth';
 import { Leaderboard } from 'pages/leaderboard';
 import { HowToModal } from 'components/how-to-modal';
 import { RouterPaths } from 'src/app.types';
