+++ packages/client/src/pages/auth/auth.module.pcss
@@ -6,7 +6,7 @@
 .panel {
   height: 100%;
   width: 100%;
-  background: rgba(255, 255, 255, 0.9);
+  background: var(--color-panel);
   border-radius: 12px;
   display: grid;
   min-width: 400px;
