+++ packages/client/src/pages/auth/index.ts
@@ -1,0 +1,1 @@
+export { RegisterPage, LoginPage } from './auth';
