--- packages/client/src/pages/register-login/index.ts
@@ -1,2 +1,0 @@
-export { RegisterPage } from './register';
-export { LoginPage } from './login';
