+++ packages/client/src/styles/variables.pcss
@@ -17,6 +17,7 @@
     #f178b6 72.13%,
     #ee8ebf 92.5%
   );
+  --color-panel: rgba(255, 255, 255, 0.9);
 
   --hover: brightness(90%);
   --font-size-base: 14px;
